refactor #101 Category remote 파일 제거
