+++ lib/features/category/data/data_sources/users_home_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/category/data/models/remote/response/users_home_res.dart';
+import 'package:debateseason_frontend_v1/features/category/data/models/response/users_home_res.dart';
 import 'package:debateseason_frontend_v1/utils/base/base_res.dart';
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
