+++ lib/features/category/data/mapper/category_mapper.dart
@@ -1,4 +1,4 @@
-import 'package:debateseason_frontend_v1/features/category/data/models/remote/response/users_home_res.dart';
+import 'package:debateseason_frontend_v1/features/category/data/models/response/users_home_res.dart';
 import 'package:debateseason_frontend_v1/features/category/domain/entities/category_entity.dart';
 
 class CategoryMapper {
