+++ lib/features/category/data/repository_impls/users_home_repository_impl.dart
@@ -1,7 +1,7 @@
-import 'package:debateseason_frontend_v1/features/category/data/data_sources/remote/users_home_data_source.dart';
+import 'package:debateseason_frontend_v1/features/category/data/data_sources/users_home_data_source.dart';
 import 'package:debateseason_frontend_v1/features/category/data/mapper/category_mapper.dart';
 import 'package:debateseason_frontend_v1/features/category/domain/entities/category_entity.dart';
-import 'package:debateseason_frontend_v1/features/category/domain/repositories/remote/users_home_repository.dart';
+import 'package:debateseason_frontend_v1/features/category/domain/repositories/users_home_repository.dart';
 import 'package:debateseason_frontend_v1/utils/base/ui_state.dart';
 import 'package:debateseason_frontend_v1/utils/logger.dart';
